requirements.txt
redis == 3.5.3
fastapi == 0.68.2
uvicorn == 0.15.0
openapi-core == 0.14.2
openapi-spec-validator == 0.3.1
httpx == 0.19.0
qwhale-client == 0.1.20
python-dotenv == 0.19.0
